% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapr_prob.R
\name{betapr_flexprob}
\alias{betapr_flexprob}
\title{Conditional probability for Beta Prime distribution}
\usage{
betapr_flexprob(K, model, features, visualise = TRUE, draws = 5, xlim)
}
\arguments{
\item{K}{Value for which P(Y<k | X) is computed.}

\item{model}{An object of class "mle2" produced using the function betapr_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional probability should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{visualise}{Logical. If TRUE (the default) the conditional distribution is plotted at P(Y<k | x) is shaded.}

\item{xlim}{Numeric vectors of length 2, giving the coordinate range of the dependent variable.}
}
\description{
betapr_flexprob returns the conditional probability P(Y<k | X) of a model fitted via the function betapr_flexfit; where α has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. The function includes a procedure for visualizing the conditional probability. betapr_flexprob also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the two parameter parametrization of the Beta Prime distribution is used in Johnson and Kotz (1995). The tow parameter distribution ins a special case of the three parameter distribution, with σ = 1. The probability probability density function is used is:

f(y) = [y^α-1 (1+y)^-(α+β)]/Β(α,β)

The function returns:

P(Y<k | X) = I(α,β)

α may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Johnson, N.L., Kotz, S., Balakrishnan, N. (1995). Continuous Univariate Distributions, Volume 2 (2nd Edition), Wiley.
}
