% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betapr_mom.R
\name{betapr_shape1}
\alias{betapr_shape1}
\title{First stage estimate for alpha in the Beta Prime distribution}
\usage{
betapr_shape1(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Beta Prime distribution.}
}
\description{
Computes the method of moments estimate for the first shape parameter in the Beta Prime dustribution.
}
