% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagum_predict.R
\name{dagum_flexpredict}
\alias{dagum_flexpredict}
\title{Conditional mean for Dagum distribution}
\usage{
dagum_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function dagum_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
dagum_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function dagum_flexfit; where b has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. dagum_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the same parametrization of the Dagum distribution as is used in Kleiber and Kotz (2003).  The probability probability density function is used is:

f(y) = apy^ap-1/[b^ap(1+(y/b)^a)^p+1]

The function returns:

E(Y|X) = bΓ(p+1/a)Γ(1-1/a)/Γ(p)

b may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003.
}
