% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_flexfit.R
\name{exp_flexfit}
\alias{exp_flexfit}
\title{Fitting Exponential distribution via maximum likelihood}
\usage{
exp_flexfit(formula, data = NULL, weights = NULL, subset = NULL,
  ownstart = NULL, key = TRUE, warnings = FALSE, ...)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which exp_flexfit is called.}

\item{weights}{An optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{ownstart}{An optional list containing starting values for the maximum likelihood estimation procedure. If a model with an intercept has been specified, the list must be of the form ownstart = list(beta1 = , …); if a mode with no intercept has been specified the list must be of the form ownstart = list(beta0 = , beta1 = , …); if the rate parameter is not a function of covariates the list must be of the form ownstart = list(lambda = ). It is important that the list have as many elements as there are parameters in the model, and that these be supplied in the order set out above.}

\item{key}{A logical parameter dictating whether a key is produced alongside the model’s output.}

\item{warnings}{A logical parameter dictating whether warnings from the maximum likelihood estimation procedure are produced alongside the model’s output.}

\item{...}{Additional arguments to be passed to the function optim within the maximum likelihood estimation procedure. Useful arguments include the gradient descent algorithm to be used and bounds on parameter values; see the stats package.}
}
\description{
exp_flexfit is used to fit an exponential distribution to a strictly positive response variable. The rate parameter may be specified either as a function of covariates or as a constant estimated using the response variable alone.
If the rate parameter is specified to be a function of covariates, the canonical log link function is used.
}
\details{
This function uses the most common parametrization of the Exponential distribution. The starting value for the maximum likelihood estimation procedure is given by the canonical estimator for the rate parameter in an Exponentially distributed random variable: the inverse of the sample mean. The probability probability density function is used is:

f(y) = λexp(-λy)

When the argument formula specifies a full model with an intercept, the rate parameter takes the following form and is estimated via a two step (least squares and maximum likelihood) procedure:

λ = exp(β0 + β1x1 + …. + βkxk)

When the formula argument specifies a model without an intercept, the rate parameter takes the bellow form and is estimated via a two step (least squares and maximum likelihood) procedure. Unless theory suggests that an intercept should not be used, users are advised to use a model with an intercept as the maximum likelihood estimation procedure is more stable.

λ = exp(β1x1 + …. + βkxk)

When a null model is specified (formula = y ~ 0) the rate parameter is not estimated as a function of covariates. The starting value for the maximum likelihood estimation procedure is obtained by calling the function exp_rate.
}
