% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_predict.R
\name{exp_flexpredict}
\alias{exp_flexpredict}
\title{Conditional mean for Exponential distribution}
\usage{
exp_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function exp_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
exp_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function exp_flexfit; where λ has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. exp_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the most common parametrization of the Exponential distribution. The probability probability density function is used is:

f(y) = λexp(-λy)

The function returns:

E(Y|X) = λ^{-1}

λ may be a function of covariates; in which case, the canonical log link function is used.
}
