% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exp_mom.R
\name{exp_rate}
\alias{exp_rate}
\title{First stage estimate for lambda in the Exponential distribution}
\usage{
exp_rate(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from an Exponential distribution.}
}
\description{
Computes the the method of moments estimate for the rate parameter in the Exponential distribution.
}
