% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisk_predict.R
\name{fisk_flexpredict}
\alias{fisk_flexpredict}
\title{Conditional mean for Fisk distribution}
\usage{
fisk_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function fisk_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
fisk_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function fisk_flexfit; where α has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. fisk_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the same parametrization of the Fisk distribution as is used in Kleiber and Kotz (2003). The probability probability density function is used is:

f(y) = αy^α-1/[β^α(1+(y/β)^α]^2

The function returns:

E(Y|X) = απ/sin(π/β)

α may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003.
}
