% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisk_mom.R
\name{fisk_scale}
\alias{fisk_scale}
\title{First stage estimate for alpha in the Fisk distribution}
\usage{
fisk_scale(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Fisk distribution.}
}
\description{
Computes the method of moments estimate for the scale parameter in the Fisk dustribution.
}
\references{
"Solving Log-Logistic Distribution Parameters From Moments." Cross Validate. N.p., 2017. Web. 11 Aug. 2018.
}
