% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisk_mom.R
\name{fisk_shape}
\alias{fisk_shape}
\title{First stage estimate for beta in the Fisk distribution}
\usage{
fisk_shape(x, lower_bound = 0.001, upper_bound = pi/2)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Fisk distribution.}

\item{lower_bound}{Lower bound in solution pace of the trigonometric equation used to find the value of beta. Ideally zero since as the solution will always be between zero and pi/2, however since the function must be defined at both endpoints zero cannot be used. By default the parameter is set to 0.001.}

\item{upper_bound}{Upper bound in solution pace of the trigonometric equation used to find the value of beta.  By default the parameter is set to pi/2.}
}
\description{
Computes the method of moments estimate for the shape parameter in the Fisk dustribution.
}
\references{
"Solving Log-Logistic Distribution Parameters From Moments." Cross Validate. N.p., 2017. Web. 11 Aug. 2018.
}
