% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_predict.R
\name{gamma_flexpredict}
\alias{gamma_flexpredict}
\title{Conditional mean for Gamma distribution}
\usage{
gamma_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function gamma_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
gamma_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function gamma_flexfit; where α has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. gamma_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the the most common parametrization of the Gamma distribution.The probability probability density function is used is:

f(y) = λ^α/Γ(α)•y^α-1 exp(-λy)

The function returns:

E(Y|X) = α/λ

α may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kempthorne. "Parameter Estimation Fitting Probability Distributions Method Of Moments." MIT 18.443 (2015)
}
