% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_mom.R
\name{gamma_rate}
\alias{gamma_rate}
\title{First stage estimate for beta in the Gamma distribution}
\usage{
gamma_rate(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Gamma distribution.}
}
\description{
Computes the method of moments estimate for the rate parameter in Gamma distribution.
}
