% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_mom.R
\name{gamma_shape}
\alias{gamma_shape}
\title{First stage estimate for alpha in the Gamma distribution}
\usage{
gamma_shape(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Gamma distribution.}
}
\description{
Computes the method of moments estimate for the shape parameter in Gamma distribution.
}
