% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb2_predict.R
\name{gb2_flexpredict}
\alias{gb2_flexpredict}
\title{Conditional mean for Generalized Beta distribution of the Second Kind}
\usage{
gb2_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function gb2_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
gb2_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function gb2_flexfit; where b has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. gb2_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the same parametrization of the Generalized Beta Distribution of the Second Kind as is used in Kleiber and Kotz (2003).  The probability probability density function is used is:

f(y) = ay^ap-1/[b^apΒ(p,q)(1+(y/b)^a)^p+q]

The function returns:

E(Y|X) = bΓ(p+1/a)Γ(q-1/a)/[Γ(p)Γ(q)]

b may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003.
}
