% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ggamma_mom.R
\name{ggamma_alpha}
\alias{ggamma_alpha}
\title{First stage estimate for alpha in the Generalized Gamma distribution}
\usage{
ggamma_alpha(x, lower = 0, upper = 100)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Generalized Gamma distribution.}

\item{lower}{Lower bound of the solution space in which the estimate may lie. By default the parameter is set to zero, since alpha is strictly positive.}

\item{upper}{Upper bound of the solution space in which the estimate may lie. By default the parameter is set to 100, although this choice is somewhat arbitrary as the estimate is generally small and the function defining the parameter space may only have one root.}
}
\description{
Computes the method of moments estimate for the first shape parameter in the Generalized Gamma dustribution. The parameterization proposed by Stacey (1962) is assumed.
}
\references{
Alberto Achcar, Jorge, Pedro Luiz Ramos, and Edson Zangiacomi Martinez. "Some Computational Aspects To Find Accurate Estimates For The Parameters Of The Generalized Gamma Distribution." Pesquisa Operacional Vol.37(2) (2017): n. pag. Print.
}
