% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma_predict.R
\name{ggamma_flexpredict}
\alias{ggamma_flexpredict}
\title{Conditional mean for Generalized Gamma distribution}
\usage{
ggamma_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function ggamma_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
ggamma_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function ggamma_flexfit; where μ has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. ggamma_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the same parametrization of the Fisk distribution as is used in Stacy (1962); note that this is differs significantly to the parametrization used in many common R packages. The probability probability density function is used is:

f(y) = [α/Γ(φ)] • μ^αφ y^αφ-1 exp(-(μy)^α)

The function returns:

E(Y|X) = μΓ(φ+α^{-1})/Γ(φ)

μ may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003.
}
