% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ggamma_mom.R
\name{ggamma_phi}
\alias{ggamma_phi}
\title{First stage estimate for phi in the Generalized Gamma distribution}
\usage{
ggamma_phi(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Generalized Gamma distribution.}
}
\description{
Computes the method of moments estimate for the second shape parameter in the Generalized Gamma dustribution. The parameterization proposed by Stacey (1962) is assumed.
}
\references{
Alberto Achcar, Jorge, Pedro Luiz Ramos, and Edson Zangiacomi Martinez. "Some Computational Aspects To Find Accurate Estimates For The Parameters Of The Generalized Gamma Distribution." Pesquisa Operacional Vol.37(2) (2017): n. pag. Print.
}
