% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_flexfit.R
\name{lnorm_flexfit}
\alias{lnorm_flexfit}
\title{Fitting Log-Normal distribution via maximum likelihood}
\usage{
lnorm_flexfit(formula, data = NULL, weights = NULL, subset = NULL,
  ownstart = NULL, key = TRUE, warnings = FALSE, ...)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lnorm_flexfit is called.}

\item{weights}{An optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{ownstart}{An optional list containing starting values for the maximum likelihood estimation procedure. If a model with an intercept has been specified, the list must be of the form ownstart = list(sigma = , beta0 = , beta1 =, …); if a mode with no intercept has been specified the list must be of the form ownstart = list(sigma = , beta1 = , …); if the rate parameter is not a function of covariates the list must be of the form ownstart = list(sigma = , mu = ). It is important that the list have as many elements as there are parameters in the model, and that these be supplied in the order set out above.}

\item{key}{A logical parameter dictating whether a key is produced alongside the model’s output.}

\item{warnings}{A logical parameter dictating whether warnings from the maximum likelihood estimation procedure are produced alongside the model’s output.}

\item{...}{Additional arguments to be passed to the function optim within the maximum likelihood estimation procedure. Useful arguments include the gradient descent algorithm to be used and bounds on parameter values; see the stats package.}
}
\description{
lnorm_flexfit is used to fit a Log-Normal distribution to a strictly positive response variable. The parameter mu may be specified either as a function of covariates or as a constant estimated using the response variable alone.
}
\details{
This function uses the two parameter parametrization of the Log-Normal distribution.Starting values for the maximum likelihood estimation procedure are taken from Brenda Faith Ginos’ Masters dissertation on the Log-Normal distribution. The probability probability density function is used is:

f(y) = [yσ(2π)^{1/2}]^{-1} exp(-log(y-μ)^{2}/(2σ^{2}))

The parameters μ and σ are not strictly location and scale parameters. Rather, they are the scale and location parameters for the normally distributed logarithm log(y), where y follows a Log-Normal distribution. In the first stage estimation functions  μ and σ are referred to as ‘effective’ scale and location parameters.

When the argument formula specifies a full model with an intercept, μ takes the following form and is estimated via a two step (least squares and maximum likelihood) procedure:

μ = β0 + β1x1 + …. + βkxk

When the formula argument specifies a model without an intercept, μ takes the bellow form and is estimated via a two step (least squares and maximum likelihood) procedure. Unless theory suggests that an intercept should not be used, users are advised to use a model with an intercept as the maximum likelihood estimation procedure is more stable.

μ = β1x1 + …. + βkxk

When a null model is specified (formula = y ~ 0) μ is not estimated as a function of covariates. The starting value for the maximum likelihood estimation procedure is obtained by calling the function lnorm_mean.
}
\references{
Faith Ginos, Brenda. "Parameter Estimation For The Lognormal Distribution." Brigham Young University Scholars Archive (2018): 1-111. Web. 10 Aug. 2018.

Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003. Print.
}
