% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_predict.R
\name{lnorm_flexpredict}
\alias{lnorm_flexpredict}
\title{Conditional mean for Log-Normal distribution}
\usage{
lnorm_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function lnorm_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
lnorm_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function lnorm_flexfit; where μ has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. lnorm_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the two parameter parametrization of the Log-Normal distribution. The probability probability density function is used is:

f(y) = [yσ(2π)^1/2]^-1 exp(-log(y-μ)^2/(2σ^2))

The function returns:

E(Y|X) = exp(μ + 0.5σ^{2})

μ may be a function of covariates; in which case, the identity link function is used.
}
\references{
Faith Ginos, Brenda. "Parameter Estimation For The Lognormal Distribution." Brigham Young University Scholars Archive (2018): 1-111. Web. 10 Aug. 2018.

Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003. Print.
}
