% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lnorm_mom.R
\name{lnorm_var}
\alias{lnorm_var}
\title{First stage estimate for sigma squared in the Log-Normal distribution}
\usage{
lnorm_var(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Log-Normal distribution.}
}
\description{
Computes the method of moments estimate for the effective dispersion parameter in the Log-Normal distru.
}
\references{
Faith Ginos, Brenda. "Parameter Estimation For The Lognormal Distribution." Brigham Young University Scholars Archive (2018): 1-111. Web. 10 Aug. 2018.
}
