% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax_predict.R
\name{lomax_flexpredict}
\alias{lomax_flexpredict}
\title{Conditional mean for Lomax distribution}
\usage{
lomax_flexpredict(model, features, draws = 5)
}
\arguments{
\item{model}{An object of class "mle2" produced using the function lomax_flexfit.}

\item{features}{A numeric vector specifying the value of covriates at which the conditional mean should be evaluated; the covariates in the vector should appear in the same order as they do in the model. Where a model does not depend on covariates the argument may be left blank.}

\item{draws}{The number of random draws from multivariate random normal representing correlated parameters. If parameter correlation is not required draws should be set to zero.}
}
\description{
lomax_flexpredict returns the conditional mean E(Y|X) of a model fitted via the function lomax_flexfit; where λ has been specified to be a function of covariates the required value should be specified using the ‘features’ parameter. lomax_flexpredict also allows for the correlation of estimated parameters via the Cholesky decomposition of the variance-covariance matrix.
}
\details{
This function uses the same parametrization of the Lomax distribution as is used in Kleiber and Kotz (2003). The probability probability density function is used is:

f(y) = (α/λ) [1 + (x/λ)]^-(α+1)

The function returns:

E(Y|X) = λ/(α-1)

λ may be a function of covariates; in which case, the cannonical log link function is used.
}
\references{
Kleiber, Christian, and Samuel Kotz. Statistical Size Distributions In Economics And Actuarial Sciences. pp. 107-147. John Wiley & Sons, 2003. Print.
}
