% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lomax_mom.R
\name{lomax_scale}
\alias{lomax_scale}
\title{First stage estimate for lambda in the Lomax distribution}
\usage{
lomax_scale(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Lomax distribution.}
}
\description{
Computes the method of moments estimate for the scale parameter in the Lomax distribution.
}
