% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lomax_mom.R
\name{lomax_shape}
\alias{lomax_shape}
\title{First stage estimate for alpha in the Lomax distribution}
\usage{
lomax_shape(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Lomax distribution.}
}
\description{
Computes the method of moments estimate for the shape parameter in the Lomax distribution.
}
