% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull_mom.R
\name{weibull_scale}
\alias{weibull_scale}
\title{First stage estimate for lambda in the Weibull distribution}
\usage{
weibull_scale(x)
}
\arguments{
\item{x}{A numeric vector of i.i.d. observations presumed to be draws from a Weibull distribution.}
}
\description{
Computes the method of moments estimate for the scale parameter in the Weibull dustribution.
}
\references{
NIOLA, VINCENZO, ROSARIO OLIVIERO, and GIUSEPPE QUAREMBA. "A Method Of Moments For The Estimation Of Weibull Pdf Parameters." Proceedings of the 8th WSEAS Int. Conference on Automatic Control, Modeling and Simulation 8 (2014)
}
