% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_select.R
\name{flex_select}
\alias{flex_select}
\title{Flex Select}
\usage{
flex_select(formula, data = NULL, IC = "AIC",
  distributions = c("exp", "lnorm", "gamma", "weibull", "lomax", "fisk",
  "ggamma", "betapr", "sinmad", "dagum", "gb2"))
}
\description{
To be written :(
}
