% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb2_flexfit.R
\name{gb2_flexfit}
\alias{gb2_flexfit}
\title{Fitting Generalized Beta of the Second Kind via maximum likelihood}
\usage{
gb2_flexfit(formula, data = NULL, weights = NULL, subset = NULL,
  ownstart = NULL, key = TRUE, warnings = FALSE, warnings1 = FALSE,
  ...)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which gamma_flexfit is called.}

\item{weights}{An optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{ownstart}{An optional list containing starting values for the maximum likelihood estimation procedure. If a model with an intercept has been specified, the list must be of the form ownstart = list(a = , p = , q = , beta0 = , beta1 =, …); if a mode with no intercept has been specified the list must be of the form ownstart = list(a = , p = , q = , beta1 = , …); if the rate parameter is not a function of covariates the list must be of the form ownstart = list(b = , a = , p = , q =). It is important that the list have as many elements as there are parameters in the model, and that these be supplied in the order set out above.}

\item{key}{A logical parameter dictating whether a key is produced alongside the model’s output.}

\item{warnings}{A logical parameter dictating whether warnings from the maximum likelihood estimation procedure are produced alongside the model’s output.}

\item{...}{Additional arguments to be passed to the function optim within the maximum likelihood estimation procedure. Useful arguments include the gradient descent algorithm to be used and bounds on parameter values; see the stats package.}
}
\description{
gb2_flexfit is used to fit a Generalized Beta distribution of the Second Kind distribution to a strictly positive response variable. The scale parameter may be specified either as a function of covariates or as a constant estimated using the response variable alone.
}
