% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_IC.R
\name{which_IC}
\alias{which_IC}
\title{General Purpose Information Criterion Function}
\usage{
which_IC(data, model, IC)
}
\arguments{
\item{model}{A fitted model object for which there exists a logLik method to extract the corresponding log-likelihood, or an object inheriting from class logLik.}

\item{IC}{...}
}
\description{
General purpose function for computing one of six preset information criteria.
}
